+++ __tests__/pages/ChoeurDesPaysduMontBlanc.test.tsx
@@ -1,7 +1,7 @@
 import { getTranslations } from 'next-intl/server';
 import { vi } from 'vitest';
 
-import MComperat, { generateMetadata } from '@/app/[locale]/projects/mcomperat/page';
+import ChoeurDesPaysduMontBlanc, { generateMetadata } from '@/app/[locale]/projects/choeurdespaysdumontblanc/page';
 
 import { render } from '../test-utils';
 
@@ -11,13 +11,13 @@ vi.mock('next-intl/server', () => ({
 
 const mockGetTranslations = vi.mocked(getTranslations);
 
-describe('MComperat Page', () => {
+describe('ChoeurDesPaysduMontBlanc Page', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it('should render the mcomperat page correctly', () => {
-    const { container } = render(<MComperat />);
+  it('should render the choeur des pays du mont blanc page correctly', () => {
+    const { container } = render(<ChoeurDesPaysduMontBlanc />);
 
     expect(container).toMatchSnapshot();
   });
@@ -25,20 +25,20 @@ describe('MComperat Page', () => {
   it('should generate metadata correctly', async () => {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const mockT = vi.fn() as any;
-    mockT.mockReturnValueOnce('mcomper.at');
-    mockT.mockReturnValueOnce('Modern, responsive web CV of Matthieu Compérat');
+    mockT.mockReturnValueOnce('Chœur des Pays du Mont Blanc');
+    mockT.mockReturnValueOnce('Website for the amateur choir based in Haute-Savoie, France');
 
     mockGetTranslations.mockResolvedValue(mockT);
 
     const params = Promise.resolve({ locale: 'en' });
     const metadata = await generateMetadata({ params });
 
-    expect(mockGetTranslations).toHaveBeenCalledWith({ locale: 'en', namespace: 'mcomperat' });
+    expect(mockGetTranslations).toHaveBeenCalledWith({ locale: 'en', namespace: 'choeurDesPaysduMontBlanc' });
     expect(mockT).toHaveBeenCalledWith('title');
     expect(mockT).toHaveBeenCalledWith('subtitle');
     expect(metadata).toEqual({
-      title: 'mcomper.at',
-      description: 'Modern, responsive web CV of Matthieu Compérat',
+      title: 'Chœur des Pays du Mont Blanc',
+      description: 'Website for the amateur choir based in Haute-Savoie, France',
     });
   });
 });